+++ src/index.ts
@@ -1,1 +1,3 @@
+export * from './utils';
+export * from './types';
 export * from './AmplienceImageStudio';
