+++ src/types/Message.types.ts
@@ -1,0 +1,15 @@
+export interface SDKEvent {
+    extensionMeta?: {
+      exportContext: string;
+    };
+    inputImageUrl?: string;
+    focus?: boolean;
+  }
+  
+  export interface ImageStudioEvent {
+    data: {
+      exportImageUrl?: string;
+      connect?: boolean;
+      disconnect?: boolean;
+    };
+  }
