+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './Message.types'
+export * from './ImageStudio.types'
